requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9

# LangChain & LLM
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-openai==0.0.5
langchain-pinecone==0.0.3

# Anthropic & OpenAI
anthropic==0.18.1
openai==1.12.0

# Vector DB & Cache
pinecone-client==3.0.0
redis==5.0.1

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
tiktoken==0.5.2

# Admin Panel
jinja2==3.1.2
aiofiles==23.2.1

# Utils
httpx==0.26.0
tenacity==8.2.3

# Security
passlib==1.7.4
